Don't panic on spurious connect request
Clients can crash the server by sending a connect request before the
connection is validated. This is trivial for bad actors, but can also
happen under poor network conditions if the validate request gets
retransmitted and the client disconnects before all the handshakes are
complete.

Simply convert the panic into a warning to avoid crashing the server.