Use handshake to seed server rtt estimate
On connect, the server is observed to resend many thousands of messages
per second, which is not expected. This seems to be caused by the RTT
estimate not being initialized before the connection process is complete,
resulting in the RTT estimate being 0ms until the first ping/pong round
trip completes.

Use the time between handshake messages to estimate RTT before the
connection is established to avoid excessive message resends on connect.