Remove zstd support
The current packet-level implementation of compression isn't very
effective. At best, it could save a bit of bandwidth, but it has no
affect on the number of packets or messages being sent. It also
complicates adding encryption.

Simply remove compression support for now. In the future, we might add
it back at the message level (so we can fit more messages per packet),
for packets that need to be fragmented, etc. At that time, we probably
don't want it to be an optional feature either.